Use single BringToForeground across the solution
