+++ src/modules/Hosts/Hosts/HostsXAML/MainWindow.xaml.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using Hosts.Helpers;
+using ManagedCommon;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Media;
 using WinUIEx;
@@ -25,7 +26,7 @@ namespace Hosts
             AppTitleTextBlock.Text = title;
 
             var handle = this.GetWindowHandle();
-            ManagedCommon.WindowHelpers.BringToForeground(handle);
+            WindowHelpers.BringToForeground(handle);
 
             Activated += MainWindow_Activated;
         }
