+++ src/modules/peek/Peek.UI/NativeMethods.txt
@@ -4,7 +4,6 @@ GetDpiForWindow
 GetForegroundWindow
 SetForegroundWindow
 SetActiveWindow
-SendInput
 GetWindowThreadProcessId
 GetCurrentThreadId
 AttachThreadInput
