+++ src/modules/peek/Peek.UI/PeekXAML/MainWindow.xaml.cs
@@ -179,7 +179,7 @@ namespace Peek.UI
             }
 
             this.Show();
-            this.BringToForeground();
+            WindowHelpers.BringToForeground(this.GetWindowHandle());
         }
 
         private Size GetMonitorMaxContentSize(Size monitorSize, double scaling)
