+++ src/settings-ui/Settings.UI/SettingsXAML/App.xaml.cs
@@ -170,7 +170,7 @@ namespace Microsoft.PowerToys.Settings.UI
 
                     // https://github.com/microsoft/microsoft-ui-xaml/issues/7595 - Activate doesn't bring window to the foreground
                     // Need to call SetForegroundWindow to actually gain focus.
-                    Utils.BecomeForegroundWindow(settingsWindow.GetWindowHandle());
+                    WindowHelpers.BringToForeground(settingsWindow.GetWindowHandle());
                 }
                 else
                 {
