+++ src/settings-ui/Settings.UI/SettingsXAML/MainWindow.xaml.cs
@@ -202,7 +202,7 @@ namespace Microsoft.PowerToys.Settings.UI
 
                     // https://github.com/microsoft/microsoft-ui-xaml/issues/7595 - Activate doesn't bring window to the foreground
                     // Need to call SetForegroundWindow to actually gain focus.
-                    Utils.BecomeForegroundWindow(flyout.GetWindowHandle());
+                    WindowHelpers.BringToForeground(flyout.GetWindowHandle());
                 });
             });
